backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
pymongo==4.10.1

# Authentication & Security
PyJWT==2.10.1
pwdlib[argon2]==0.2.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Environment variables
python-dotenv==1.0.1
